chore:  merge token deployment step 2 into step 1
