--- apps/maestro/src/features/InterchainTokenDeployment/steps/token-settings/index.ts
@@ -1,2 +1,0 @@
-export * from "./TokenSettings";
-export { default } from "./TokenSettings";
